+++ gui/src/components/settings/SettingsLayout.tsx
@@ -40,6 +40,10 @@ export function SettingSelectorMobile() {
         label: l10n.getString('settings-sidebar-serial'),
         value: { url: '/settings/serial' },
       },
+      {
+        label: l10n.getString('settings-sidebar-advanced'),
+        value: { url: '/settings/advanced' },
+      },
     ];
 
   const { control, watch, handleSubmit, setValue } = useForm<{
