+++ gui/src/components/settings/SettingsResetModal.tsx
@@ -35,16 +35,16 @@ export function SettingsResetModal({
     >
       <div className="flex w-full h-full flex-col ">
         <div className="flex flex-col flex-grow items-center gap-3">
-          <Localized id="settings-reset_warning" elems={{ b: <b></b> }}>
+          <Localized id="settings-utils-advanced-reset_warning" elems={{ b: <b></b> }}>
             <WarningBox>
-              <b>Warning:</b> This will reset all your settings to the defaults.
+              <b>Warning:</b> This will reset all your settings to the defaults.<br />
               Are you sure you want to do this?
             </WarningBox>
           </Localized>
 
           <div className="flex flex-row gap-3 pt-5 place-content-center">
             <Button variant="primary" onClick={onClose}>
-              {l10n.getString('settings-reset_warning-cancel')}
+              {l10n.getString('settings-utils-advanced-reset_warning-cancel')}
             </Button>
             <Button
               variant="tertiary"
@@ -52,7 +52,7 @@ export function SettingsResetModal({
                 accept();
               }}
             >
-              {l10n.getString('settings-reset_warning-reset')}
+              {l10n.getString('settings-utils-advanced-reset_warning-reset')}
             </Button>
           </div>
         </div>
